requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.1.5
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-js-asset==1.2.2
django-meta==1.4.1
django-seo==0.3.5
django-suit-ckeditor==0.0.2
djangorestframework==3.9.1
DjangoSEO==1.0
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
pipenv==2018.11.26
pycodestyle==2.5.0
pylint==2.2.2
python-dateutil==2.7.5
pythonanywhere==0.7.4
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
virtualenv==16.2.0
virtualenv-clone==0.5.1
virtualenvwrapper-win==1.2.5
whitenoise==4.1.2
wrapt==1.11.0